backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
opencv-python==4.5.3.56
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
ultralytics==0.0.0
pyttsx3==2.90
playsound==1.2.2